% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{abs_diff_lt}
\alias{abs_diff_lt}
\title{test difference below a threshold}
\usage{
abs_diff_lt(y, z, abs_diff = 0.05)
}
\arguments{
\item{y}{variable 1}

\item{z}{variable 2}

\item{abs_diff}{absolute difference to test in percent}
}
\description{
test difference between 2 values. The absolute value
should be below a given threshold.
}
\seealso{
\code{\link{qualcon}}
}
