% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\alias{agg_data.raw_level}
\alias{agg_data.raw_prod_type}
\alias{agg_data.raw_prod_renewable_type}
\title{Aggregate raw dataset from country rules}
\usage{
agg_data(x, ...)

\method{agg_data}{raw_level}(x, ...)

\method{agg_data}{raw_prod_type}(x, ...)

\method{agg_data}{raw_prod_renewable_type}(x, ...)
}
\arguments{
\item{x}{raw dataset.}

\item{...}{arguments to be passed to methods.}
}
\description{
From a raw dataset and a set of rules, aggregations are performed
to produce for each country, possible date points (and
eventually other dimensions) a set of measure(s).
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)


data_neg_val <- structure(list(country = rep("LUXEMBOURG", 24),
DateTime = structure(seq(1502582400, to = 1502665200, by = 3600),
class = c("POSIXct", "POSIXt")), BZN = c(346.000000000003,
352.000000000003, 354.999999999998, 330.000000000002, 345.000000000003,
363.999999999998, 367, 374,
389.999999999997, 373, 351, 342, 339, 344, 367, 386, -36032.64,
398.999999999998, 414.000000000005, 412.000000000005, 381.999999999998,
351.999999999998, 336.000000000002, 332.999999999995), CTA = c(346,
352, 355, 330, 345, 364, 367, 374, 390, 373, 351, 342, 339, 344,
367, 386, 401, 399, 414, 412, 382, 352, 336, 333), CTY = c(346,
352, 355, 330, 345, 364, 367, 374, 390, 373, 351, 342, 339, 344,
367, 386, 401, 399, 414, 412, 382, 352, 336, 333)), .Names = c("country",
"DateTime", "BZN", "CTA", "CTY"), row.names = 179664:179687,
class = c("data.frame",
"aggregated"))

aggregated_db <- augment_validation(data_neg_val)
aggregated_db[data.table::hour(aggregated_db$DateTime) \%in\% 18,
              "BZN_LAG_LT_30"]
}
