% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_prod_with_rules.R
\name{aggregate_prod_with_rules}
\alias{aggregate_prod_with_rules}
\title{Aggregate production raw dataset from country rules}
\usage{
aggregate_prod_with_rules(x)
}
\arguments{
\item{x}{raw dataset (an object of class \code{raw_channel_prod})}
}
\description{
From a raw dataset and a set of rules, aggregations are performed
to produce for each country and possible date points a set of three measures: CTY, CTA
and BZN.
}
