% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_with_rules.R
\name{aggregate_with_rules}
\alias{aggregate_with_rules}
\title{Aggregate raw dataset from country rules}
\usage{
aggregate_with_rules(x, file_rules = NULL)
}
\arguments{
\item{x}{raw dataset (an object of class \code{raw_level})}

\item{file_rules}{yaml file containing for each country aggregation rules}
}
\description{
From a raw dataset and a set of rules, aggregations are performed
to produce for each country and possible date points a set of three measures: CTY, CTA
and BZN.
}
\examples{
data(load_example)
cty_rules <- system.file(package = "entsoe", "templates/cty_rules.yaml")
val_rules <- system.file(package = "entsoe", "templates/raw/validate.yml")
fp_rules <- system.file(package = "entsoe", "templates/raw/false_positives.yml")
raw_db <- augment_rules(load_example, file_rules = cty_rules)
raw_db <- augment_validation(data = raw_db, val_rules = val_rules,
  fp_rules = fp_rules)

agg_data <- aggregate_with_rules(raw_db)
}
