% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_load_read.R
\name{anta_load}
\alias{anta_load}
\title{import load data from an entsoe repository}
\usage{
anta_load(data_dir = NULL)
}
\arguments{
\item{data_dir}{datasets directory}
}
\description{
import csv data representing load data
from an entsoe repository.
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")
load_data <- anta_load(data_dir = load_dir )
}
