% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_prod_filiere.R
\name{anta_prod_channel}
\alias{anta_prod_channel}
\title{import production data per channels from an entsoe repository}
\usage{
anta_prod_channel(production_dir = NULL, capacity_dir = NULL)
}
\arguments{
\item{production_dir}{datasets directory of production by type files}

\item{capacity_dir}{datasets directory for capacities by type files}
}
\description{
import csv data representing production per channels data
from an entsoe repository.
}
\examples{
production_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B01")
capacity_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B06")
raw_channel <- anta_prod_channel(production_dir, capacity_dir)
}
