% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_prod_groups.R
\name{anta_prod_group}
\alias{anta_prod_group}
\title{import production data per groups from an entsoe repository}
\usage{
anta_prod_group(production_dir = NULL)
}
\arguments{
\item{production_dir}{datasets directory of production by type files}
}
\description{
import csv data representing production per groups data
from an entsoe repository.
}
\examples{
\dontrun{
data_path <- file.path("/Users/davidgohel/Documents/consulting/RTE",
  "PROD2016/B02-production_realisee_par_groupe")
data <- anta_prod_group(production_dir = data_path)
}
}
