% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_prod_renewable.R
\name{anta_prod_renewable}
\alias{anta_prod_renewable}
\title{renewable production data per types of production}
\usage{
anta_prod_renewable(production_dir = NULL, capacity_dir = NULL)
}
\arguments{
\item{production_dir}{datasets directory of data energy productions
by types.}

\item{capacity_dir}{datasets directory of data energy capacities
by types.}
}
\description{
import csv data representing production per types. The data
is read from an entsoe repository.
}
