% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_prod_type.R
\name{anta_prod_type}
\alias{anta_prod_type}
\title{production data per types of production}
\usage{
anta_prod_type(production_dir = NULL, capacity_dir = NULL)
}
\arguments{
\item{production_dir}{datasets directory of data energy productions
by types.}

\item{capacity_dir}{datasets directory of data energy capacities
by types.}
}
\description{
import csv data representing production per types. The data
is read from an entsoe repository.
}
\examples{
production_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B01")
capacity_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B06")
prod_by_types <- anta_prod_type(production_dir, capacity_dir)
}
