% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{are_equals_if_not_na}
\alias{are_equals_if_not_na}
\title{test equality}
\usage{
are_equals_if_not_na(y, z)
}
\arguments{
\item{y}{variable 1}

\item{z}{variable 2}
}
\description{
test equality when 2 values are not NA.
This function is provided to be used in validity checking.
}
\seealso{
\code{\link{qualcon}}
}
