% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_daily_mean.R
\name{augment_daily}
\alias{augment_daily}
\title{add holiday column}
\usage{
augment_daily(x, col, country_id = "country", decay = 1)
}
\arguments{
\item{x}{dataset}

\item{col}{column name of the measure to be aggregated and lagged}

\item{country_id}{column name of the country to be used in aggregations}

\item{decay}{lag as a number of days}
}
\description{
add holiday column
}
\examples{

if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
}

if( require(magrittr) )
  agg_db <- aggregate_with_rules(load_data) \%>\%
    augment_validation() \%>\%
    data_correct_with_rules() \%>\%
    augment_process_summary() \%>\%
    augment_holiday() \%>\%
    augment_seasons_id() \%>\%
    augment_daylight() \%>\%
    augment_daily(col = "CTY", decay = 1) \%>\%
    augment_daily(col = "CTY", decay = 2)
}
