% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_daylight.R
\name{augment_daylight}
\alias{augment_daylight}
\title{add day light columns}
\usage{
augment_daylight(x, ts_key = "DateTime", loc_id = "country")
}
\arguments{
\item{x}{dataset}

\item{ts_key}{column name specifying the datetime column}

\item{loc_id}{column name specifying the location column (country)}
}
\description{
add day light columns
}
\examples{

if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
}

if( require(magrittr) )
  agg_db <- aggregate_with_rules(load_data) \%>\%
    augment_validation() \%>\%
    data_correct_with_rules() \%>\%
    augment_process_summary() \%>\%
    augment_holiday() \%>\%
    augment_seasons_id() \%>\%
    augment_daylight()
}
