% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_holiday.R
\name{augment_holiday}
\alias{augment_holiday}
\title{add holiday column}
\usage{
augment_holiday(x, country_id = "country")
}
\arguments{
\item{x}{dataset}

\item{country_id}{column name specifying the country column}
}
\description{
add holiday column
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
aggregated_db <- augment_holiday(aggregated_db)
head(aggregated_db)
}
