% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_process_summary.R
\name{augment_process_summary}
\alias{augment_process_summary}
\title{add summary columns}
\usage{
augment_process_summary(data, colname = "summary")
}
\arguments{
\item{data}{dataset}

\item{colname}{name of the column to create that will contain
processed informations (i.e. original, invalid, corrected, ...)}
}
\description{
add validation column
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
head(aggregated_db)
}
