% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_rules.R
\name{augment_rules}
\alias{augment_rules}
\title{add to a raw dataset columns relative to country aggregation}
\usage{
augment_rules(raw_db, file_rules = NULL)
}
\arguments{
\item{raw_db}{raw dataset}

\item{file_rules}{yaml file containing for each country aggregation rules}
}
\description{
add country column to a raw dataset to enable
calculations of CTY, CTA and BZN
}
\examples{
data(load_example)
cty_rules <- system.file(package = "entsoe", "templates/cty_rules.yaml")
raw_db <- augment_rules(load_example, file_rules = cty_rules)
}
