% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_seasons_id.R
\name{augment_seasons_id}
\alias{augment_seasons_id}
\title{add weekly season id columns}
\usage{
augment_seasons_id(x, ts_key = "DateTime")
}
\arguments{
\item{x}{dataset}

\item{ts_key}{column name specifying the datetime column}
}
\description{
add weekly season id columns
}
\examples{

if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
}

if( require(magrittr) )
  agg_db <- aggregate_with_rules(load_data) \%>\%
    augment_validation() \%>\%
    data_correct_with_rules() \%>\%
    augment_process_summary() \%>\%
    augment_holiday() \%>\%
    augment_seasons_id()
}
