% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_validation.R
\name{augment_validation}
\alias{augment_validation}
\title{evaluate validation rules for each line of a dataset}
\usage{
augment_validation(data, val_rules = NULL, fp_rules = NULL)
}
\arguments{
\item{data}{dataset}

\item{val_rules}{yaml file containing validation rules}

\item{fp_rules}{yaml file containing false positive rules to drop from results}
}
\description{
add logical columns to a dataset. Each column is a test to
perform against dataset.
}
\examples{
data(load_example)
cty_rules <- system.file(package = "entsoe", "templates/cty_rules.yaml")
val_rules <- system.file(package = "entsoe", "templates/raw/validate.yml")
fp_rules <- system.file(package = "entsoe", "templates/raw/false_positives.yml")
raw_db <- augment_rules(load_example, file_rules = cty_rules)
raw_db <- augment_validation(data = raw_db, val_rules = val_rules,
  fp_rules = fp_rules)
}
