% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{build_db}
\alias{build_db}
\title{tidy raw data}
\usage{
build_db(raw_db)
}
\arguments{
\item{raw_db}{data returned by \code{\link{import_load_db}}}
}
\description{
tidy data imported with function \code{import_load_db}.
A set of rules defined in cty_rules.yaml will be used to reshape untidy data
}
