% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_db.R
\name{correct_db}
\alias{correct_db}
\title{correct a data frame}
\usage{
correct_db(db, v_rules = NULL, c_rules = NULL)
}
\arguments{
\item{db}{data to be eventually modified}

\item{v_rules}{yaml file containing rules to evaluate for quality evaluation}

\item{c_rules}{yaml file containing rules for corrections}
}
\description{
Correction of a data frame with a set of yaml rules
}
