% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{country_coordinates}
\alias{country_coordinates}
\title{countries coordinates}
\format{A data frame with 3 variables, \code{country}, \code{lat} and \code{lon}.}
\usage{
data(country_coordinates)
}
\description{
The dataset describes for each country the corresponding latitude and longitude.
}
\details{
\itemize{
  \item country country name
  \item lat,lon latitude and longitude
}
}
\keyword{datasets}
