% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_correct.R
\name{data_correct}
\alias{data_correct}
\title{correct datasets}
\usage{
data_correct(data, corrections_rules = NULL)
}
\arguments{
\item{data}{dataset}

\item{corrections_rules}{yaml file containing correction rules}
}
\description{
correct data based on condition expressed in a yaml file.
}
\examples{
data(load_example)
cty_rules <- system.file(package = "entsoe",
  "templates/cty_rules.yaml")
val_rules <- system.file(package = "entsoe",
  "templates/raw/validate.yml")
fp_rules <- system.file(package = "entsoe",
  "templates/raw/false_positives.yml")
correct_rules <- system.file(package = "entsoe",
  "templates/aggregated/correct.yml")
raw_db <- augment_rules(load_example, file_rules = cty_rules)
raw_db <- augment_validation(data = raw_db, val_rules = val_rules,
  fp_rules = fp_rules)

val_rules <- system.file(package = "entsoe",
  "templates/aggregated/validation.yml")
fp_rules <- system.file(package = "entsoe",
  "templates/aggregated/false_positives.yml")
agg_data <- aggregate_with_rules(raw_db)
agg_data <- augment_validation(agg_data, val_rules = val_rules,
  fp_rules = fp_rules)
data_correct(agg_data, corrections_rules = correct_rules)
}
