% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_correct.R
\name{data_correct_with_rules}
\alias{data_correct_with_rules}
\title{correct datasets}
\usage{
data_correct_with_rules(data)
}
\arguments{
\item{data}{dataset}
}
\description{
correct data based on condition expressed in a yaml file.
}
\examples{
if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
  agg_db <- aggregate_with_rules(load_data)
  agg_db <- augment_validation(agg_db)
  agg_db <- data_correct_with_rules(agg_db)
}
}
