% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{daylight}
\alias{daylight}
\title{countries daylights}
\format{A data frame with 4 variables, \code{country}, \code{date}, \code{sunset}, \code{sunrise}.}
\usage{
data(daylight)
}
\description{
The dataset describes for each country and for 10 years the corresponding sunset and sunrise time
}
\keyword{datasets}
