% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cty_rf.R
\name{define_model_rf}
\alias{define_model_rf}
\title{define and save model}
\usage{
define_model_rf(data, x_vars, y_var, save_model_dir = getwd(), id = "",
  h2o_port = 54321, h2o_ip = "localhost")
}
\arguments{
\item{data}{aggregated load dataset}

\item{x_vars}{x variables}

\item{y_var}{y variable}

\item{save_model_dir}{directory where to save models}

\item{id}{models id, will be used to prefix files and qualify produced models}

\item{h2o_port}{port number of the H2O server}

\item{h2o_ip}{IP address of the server where H2O is running}
}
\description{
create a model for each country of an validated and corrected
load dataset. Each model is saved in a file.
}
