% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_repository.R
\name{entsoe_repository}
\alias{entsoe_repository}
\title{create an entsoe repository ready to be analysed}
\usage{
entsoe_repository(wrangling_dir, data_dir, utf16 = TRUE, force = FALSE)
}
\arguments{
\item{wrangling_dir}{workspace directory}

\item{data_dir}{datasets directory}

\item{utf16}{whether text file are encoded in UTF16 (if not, UTF8)}

\item{force}{should the directories be all initiated if already existing}
}
\description{
an entsoe repository consists in a serie of folder, each representing
a successive step in the data treatment of entsoe datasets. Each folder will contain
output datamarts, qualified invalidations and eventually reports.
}
