% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_db.R
\name{extract_nonvalid_data}
\alias{extract_nonvalid_data}
\title{eval validation rules against a dataset}
\usage{
extract_nonvalid_data(db)
}
\arguments{
\item{db}{data to be confronted with rules}
}
\description{
Confront data with a set of validation rules
}
