% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_data.R
\name{extract_raw_data}
\alias{extract_raw_data}
\title{extract raw data to be linked with errors diagnostics}
\usage{
extract_raw_data(raw_db, issues_db)
}
\arguments{
\item{raw_db}{data returned by \code{\link{read_load_files}}}

\item{issues_db}{data returned by \code{\link{qualcon}}}
}
\description{
extract raw data to be linked with errors diagnostics
}
