% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{fortify_from_rules}
\alias{fortify_from_rules}
\title{tidy raw data}
\usage{
fortify_from_rules(raw_db, file_rules = NULL)
}
\arguments{
\item{raw_db}{data returned by \code{\link{read_load_files}}}

\item{file_rules}{yaml file containing rules for each country}
}
\description{
tidy data imported with function \code{read_load_files}.
A set of rules defined in yaml_data/cty_rules.yaml will be used to reshape untidy data
}
\section{rules}{

A set of rules is to be defined for each country. Items CTY, CTA and BZN
must be provided as array of MapCode to be aggregated together.
}

\examples{
library(antadraft)
rep_path <- system.file(package = "antadraft", "files/load")
load_db <- read_load_files(rep_path)
db <- fortify_from_rules(raw_db = load_db)
}
