% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_db.R
\name{fortify_invalidated_data}
\alias{fortify_invalidated_data}
\title{eval validation rules against a dataset}
\usage{
fortify_invalidated_data(dat)
}
\arguments{
\item{db}{data to be confronted with rules}
}
\description{
Confront data with a set of validation rules
}
