% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_load_db.R
\name{fortify_load_db}
\alias{fortify_load_db}
\title{parse load data files into a tibble}
\usage{
fortify_load_db(dir_src)
}
\arguments{
\item{dir_src}{data source directory}
}
\description{
parse check and correct data to build a unique dataset containing
corrected values
}
\note{
file are supposed to be UTF-16 encoded.
}
