% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{fortify_qualcon}
\alias{fortify_qualcon}
\title{eval validation rules against a dataset}
\usage{
fortify_qualcon(dat)
}
\arguments{
\item{dat}{errors data.frame to be summarised. It should be a
data.frame returned by the call to \code{\link{qualcon}}}
}
\description{
Confront data with a set of validation rules
}
