% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULL.R
\docType{data}
\name{fr_agg}
\alias{fr_agg}
\title{aggregated example dataset}
\format{A data frame with 4 rows and 5 variables}
\usage{
data(fr_agg)
}
\description{
A dataset containing 4 lines of data, first line has a missing CTA value.
}
\details{
\itemize{
  \item country country
  \item DateTime time value
  \item BZN,CTA, CTY aggregated values
}
}
\keyword{datasets}
