% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_db.R
\name{get_invalidated_data}
\alias{get_invalidated_data}
\title{eval validation rules against a dataset}
\usage{
get_invalidated_data(db, yaml_rules = system.file(package = "antadraft",
  "validation_rules.yml"))
}
\arguments{
\item{db}{data to be confronted with rules}
}
\description{
Confront data with a set of validation rules
}
