% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{holidays}
\alias{holidays}
\title{European holidays}
\format{A data frame with 4 variables, \code{Date}, \code{country},
\code{is_off}, \code{likely_off}.}
\usage{
data(holidays)
}
\description{
A raw dataset containing for each country dates of public holidays and a
column indicating if the previous day could be used as holiday because
preceding a public holidays.
}
\keyword{datasets}
