% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_load_db.R
\name{import_load_db}
\alias{import_load_db}
\title{import load data files into a tibble}
\usage{
import_load_db(dir_src)
}
\arguments{
\item{dir_src}{data source directory}
}
\description{
import load data files into a tibble
}
\note{
file are supposed to be UTF-16 encoded.
}
