% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cty_rf.R
\name{impute_cty_rf}
\alias{impute_cty_rf}
\title{CTY imputation with random forests}
\usage{
impute_cty_rf(db, hour_decay = -1, loop = 3, h2o_port = 54321,
  h2o_ip = "localhost", save_model = FALSE, save_model_dir = getwd())
}
\arguments{
\item{db}{aggregated load dataset}

\item{hour_decay}{integer specifying to shift data Y from \code{hour_decay}
and add the resulting column as variable}

\item{loop}{specify how many correction loops have to be run}

\item{h2o_port}{port number of the H2O server}

\item{h2o_ip}{IP address of the server where H2O is running}

\item{save_model}{should the models be saved on disk}

\item{save_model_dir}{where models should be saved}
}
\description{
CTY imputation with random forests
}
\examples{
\dontrun{
load_dir <- system.file(package = "antaDraft", "data_sample")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- aggregate_with_rules(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)

corrected_by_rf <- impute_cty_rf(aggregated_db, hour_decay = -1, loop = 4)
corrected_by_rf <- impute_cty_rf(corrected_by_rf, hour_decay = 1, loop = 3)
}
}
