% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cty_rf.R
\name{impute_with_model}
\alias{impute_with_model}
\alias{update_learning_db}
\title{impute invalid data with defined model}
\usage{
impute_with_model(db, id, rmsle_max = 0.03, h2o_port = 54321,
  h2o_ip = "localhost")

update_learning_db(db, target = "CTY")
}
\arguments{
\item{db}{aggregated load dataset}

\item{id}{model id}

\item{rmsle_max}{error max of a model, if less, model will be used to replace
invalid data.}

\item{h2o_port}{port number of the H2O server}

\item{h2o_ip}{IP address of the server where H2O is running}

\item{target}{to be removed, model target}
}
\description{
impute invalid data with models created by \code{\link{define_model_rf}}
}
\examples{
\dontrun{
load_dir <- "/Users/davidgohel/Documents/consulting/RTE/load_files"

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)

dat <- as_learning_db(aggregated_db )

x_vars <- c(
  "year.iso", "week.iso", "hour.iso",
  "day.iso", "light_time", "is_off", "likely_off",
  "DAILY_MIN_CTY_MINUS_1", "DAILY_AVG_CTY_MINUS_1", "DAILY_MAX_CTY_MINUS_1",
  "HOUR_SHIFT_CTY_MINUS_1")

dat <- define_model_rf( data = dat, x_vars = x_vars, y_var = "CTY",
                        save_model_dir = file.path( getwd(), "ttt"),
                        id = "BACKWARD" )

x_vars <- c(
  "year.iso", "week.iso", "hour.iso", "day.iso", "light_time",
  "is_off", "likely_off", "DAILY_MIN_CTY_PLUS_1",
  "DAILY_AVG_CTY_PLUS_1", "DAILY_MAX_CTY_PLUS_1", "HOUR_SHIFT_CTY_PLUS_1")

dat <- define_model_rf( data = dat, x_vars = x_vars, y_var = "CTY",
                        save_model_dir = file.path( getwd(), "ttt"),
                        id = "FORWARD" )
for(i in 1:10 ){
  dat <- impute_with_model(dat, id = "FORWARD")
  dat <- impute_with_model(dat, id = "BACKWARD")
  dat <- update_learning_db(dat)
}
}
}
