% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_antares_project.R
\name{init_antares_project}
\alias{init_antares_project}
\alias{add_load_to_project}
\alias{add_wind_to_project}
\alias{add_solar_to_project}
\alias{add_ror_to_project}
\alias{add_hwr_to_project}
\title{create an empty antares project}
\usage{
init_antares_project(path)

add_load_to_project(data, start_time, end_time)

add_wind_to_project(data, start_time, end_time)

add_solar_to_project(data, start_time, end_time)

add_ror_to_project(data, start_time, end_time,
  productions = c("Hydro Pumped Storage", "Hydro Water Reservoir",
  "Hydro Run-of-river and poundage"))

add_hwr_to_project(data, start_time, end_time)
}
\arguments{
\item{path}{folder to be created}

\item{data}{dataset}

\item{start_time, end_time}{time limits (posixct)}
}
\description{
create an empty antares project to be then
enriched with data.
}
\examples{
\dontrun{
library(antaresEditObject)
library(magrittr)

load_dir <- "/.../load_20180115"
load_data <- anta_load(data_dir = load_dir ) \%>\% agg_data()

production_dir <- "/.../prod_20180115/B01"
capacity_dir <- "/.../prod_20180115/B06"

global_options <- getOption("global_options")
p_renewable_file <- global_options$renewable_production_per_country

PTT <- read_prod_type(
  production_dir = production_dir,
  capacity_dir = capacity_dir,
  production_file = p_renewable_file) \%>\%
  agg_data()

start_time <- fasttime::fastPOSIXct("2017-01-01 00:00:00", tz = "GMT")
end_time <- fasttime::fastPOSIXct("2017-01-31 23:00:00", tz = "GMT")

init_antares_project("demo_proj")
add_load_to_project(load_data,  start_time = start_time,  end_time = end_time )
add_wind_to_project(PTT, start_time = start_time, end_time = end_time )
add_solar_to_project(PTT, start_time = start_time, end_time = end_time )
add_ror_to_project(PTT, start_time = start_time, end_time = end_time )
add_hwr_to_project(PTT, start_time = start_time, end_time = end_time )
}
}
