% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{is_positive}
\alias{is_positive}
\title{test positive value}
\usage{
is_positive(y)
}
\arguments{
\item{y}{variable to test}
}
\description{
test if values are positive when not NA
}
\seealso{
\code{\link{qualcon}}
}
