% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_utils.R
\name{lag_diff_ok}
\alias{lag_diff_ok}
\title{test lag values}
\usage{
lag_diff_ok(y)
}
\arguments{
\item{y}{variable to test}
}
\description{
test if values are not less than fifty percent of their preceding values.
}
\seealso{
\code{\link{qualcon}}
}
