% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULL.R
\docType{data}
\name{load_example}
\alias{load_example}
\title{Dummy raw dataset for unit tests}
\format{A data frame with 216 rows and 9 variables}
\usage{
data(load_example)
}
\description{
A raw dataset containing few time points describing load data. To be used for
unit tests.
}
\keyword{datasets}
