% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{load_gerluxaus_20150115}
\alias{load_gerluxaus_20150115}
\title{load data for Germany, Luxembourg and Austria}
\format{A data frame with 12 sampled lines.}
\usage{
data(load_gerluxaus_20150115)
}
\description{
The dataset describes raw load data sent by entsoe.
}
\details{
\itemize{
  \item DateTime time point of measures
  \item country country name
  \item AreaTypeCode BZN, CTY or CTA
  \item TotalLoadValue load value
}
}
\keyword{datasets}
