% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.aggregated}
\alias{plot.aggregated}
\title{plot an aggregated dataset}
\usage{
\method{plot}{aggregated}(x, y = "CTY", subset = NULL, ...)
}
\arguments{
\item{x}{aggregated dataset (returned by \code{agg_data()})}

\item{y}{one of the available measure (.i.e CTY, CTA, BZN)}

\item{subset}{a logical vector used to subset x before plotting}

\item{...}{unused parameter}
}
\description{
display on a new plot aggregated time serie
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")
load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
aggregated_db <- agg_data(load_data)
plot(aggregated_db)
}
