% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.controled}
\alias{plot.controled}
\title{plot validators}
\usage{
\method{plot}{controled}(x, subset = NULL, ...)
}
\arguments{
\item{x}{controled dataset (returned by \code{augment_validation()})}

\item{subset}{a logical vector used to subset x before plotting}

\item{...}{parameters to be sent to \code{\link[UpSetR]{upset}}}
}
\description{
display on a new plot validators grouped by intersections.
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")
load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
plot( aggregated_db, nsets = 7 )
}
