% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.qualcon_agg}
\alias{plot.qualcon_agg}
\alias{plot.qualcon_raw}
\title{plot qualcon object}
\usage{
\method{plot}{qualcon_agg}(x, subset = NULL, ...)

\method{plot}{qualcon_raw}(x, subset = NULL, ...)
}
\arguments{
\item{x}{qualcon object}

\item{subset}{a logical vector used to subset x before plotting}

\item{...}{unused parameter}
}
\description{
display on a new plot qualcon informations
}
\examples{
load_dir <- system.file(package = "antaDraft", "data_sample")
load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(load_data)

raw_qualcon <- qualcon(load_data)
plot(raw_qualcon)

aggregated_db <- aggregate_with_rules(load_data)
aggregated_db <- augment_validation(aggregated_db)

add_qualcon <- qualcon(aggregated_db)
plot(add_qualcon)
}
