% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_validity.R
\name{pre_qualcon}
\alias{pre_qualcon}
\title{eval validation rules against a raw dataset}
\usage{
pre_qualcon(db, rules = system.file(package = "antadraft",
  "yaml_data/validate/pre_validation_rules.yml"))
}
\arguments{
\item{db}{data to be confronted with rules}

\item{rules}{yaml file containing rules to be used to validate data}
}
\description{
Confront raw data with a set of validation rules
}
