% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_compare.R
\name{prod_compare}
\alias{prod_compare}
\title{compare channel and group production data}
\usage{
prod_compare(channel_prod, group_prod)
}
\arguments{
\item{channel_prod}{data produced by \code{anta_prod_channel}}

\item{group_prod}{data produced by \code{anta_prod_group}}
}
\description{
compare channel and group production data.
}
