% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{qualcon}
\alias{qualcon}
\alias{qualcon.raw_level}
\alias{qualcon.aggregated}
\alias{qualcon.type_group_prod}
\alias{qualcon.raw_prod_type}
\alias{qualcon.raw_prod_renewable_type}
\alias{qualcon.raw_prod_group}
\alias{qualcon.aggregated_prod}
\alias{qualcon.aggregated_renewable_prod}
\alias{aggregated_renewable_prod}
\title{perform quality control}
\usage{
qualcon(x)

\method{qualcon}{raw_level}(x)

\method{qualcon}{aggregated}(x)

\method{qualcon}{type_group_prod}(x)

\method{qualcon}{raw_prod_type}(x)

\method{qualcon}{raw_prod_renewable_type}(x)

\method{qualcon}{raw_prod_group}(x)

\method{qualcon}{aggregated_prod}(x)

\method{qualcon}{aggregated_renewable_prod}(x)

aggregated_renewable_prod(x)
}
\arguments{
\item{x}{data where controls have been evaluated}
}
\description{
perform quality control
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")
load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(load_data)

qualcon(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)

qualcon(aggregated_db)
}
