% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_load_db.R
\name{read_load_files}
\alias{read_load_files}
\title{import load data}
\usage{
read_load_files(dir_src)
}
\arguments{
\item{dir_src}{data source directory}
}
\description{
import all load data csv files from a directory.
}
\section{file encoding}{

file are expected to be UTF-16 encoded. This should be changed to UTF8 as
it increases file size and is not a standard file encoding and is likely to
cause troubles in the future.
}

\examples{
\donttest{
library(antadraft)
library(magrittr)

rep_path <- "D:/transparency_repo/staging_area/incoming/A-CONSOMMATION/A01-Consommation_réalisée"
load_db <- rep_path \%>\% read_load_files()
}
}
