% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prod_type.R
\name{read_prod_type}
\alias{read_prod_type}
\title{production data per types of production}
\usage{
read_prod_type(production_dir = NULL, capacity_dir = NULL,
  production_file = NULL)
}
\arguments{
\item{production_dir}{datasets directory of data energy productions
by types.}

\item{capacity_dir}{datasets directory of data energy capacities
by types.}

\item{production_file}{YAML production file to be used. This file
should contain for each country a list of production types.}
}
\description{
import csv data representing production per types. The data
is read from an entsoe repository.
}
\examples{
production_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B01")
capacity_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B06")
global_options <- getOption("global_options")
prod_by_types <- read_prod_type(production_dir = production_dir,
  capacity_dir = capacity_dir,
  production_file = global_options$thermal_production_per_country)

# Existing lists of production types -----
system.file(package = "antaDraft", "config/global/thermal_production.yml")
system.file(package = "antaDraft", "config/global/hps_production.yml")
system.file(package = "antaDraft", "config/global/thermal_production.yml")
}
