% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULL.R
\docType{data}
\name{sample_composite}
\alias{sample_composite}
\title{Dummy raw dataset for unit tests}
\format{A data frame with 12 rows and 11 variables}
\usage{
data(sample_composite)
}
\description{
A raw dataset containing few time points describing load data. To be used for
unit tests.
}
\keyword{datasets}
