% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{set_antadraft_load_option}
\alias{set_antadraft_load_option}
\title{set options for load data processing}
\usage{
set_antadraft_load_option(cty_rules = NULL, raw_val = NULL,
  raw_val_fp = NULL, agg_val = NULL, agg_val_fp = NULL,
  correct_rules = NULL)
}
\arguments{
\item{cty_rules}{rules used when aggregating data}

\item{raw_val}{rules to validate raw data}

\item{raw_val_fp}{rules to exclude irrelevant raw validations because defined as false positive}

\item{agg_val}{rules to validate aggregated data}

\item{agg_val_fp}{rules to exclude irrelevant aggregation validations because defined as false positive}
}
\description{
set options for load data processing by specifying
\code{yaml} files to be used at various stages.
}
