% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\alias{agg_data.raw_level}
\alias{agg_data.raw_prod_type}
\alias{agg_data.raw_prod_renewable_type}
\title{Aggregate raw dataset from country rules}
\usage{
agg_data(x, ...)

\method{agg_data}{raw_level}(x, ...)

\method{agg_data}{raw_prod_type}(x, ...)

\method{agg_data}{raw_prod_renewable_type}(x, ...)
}
\arguments{
\item{x}{raw dataset.}

\item{...}{arguments to be passed to methods.}
}
\description{
From a raw dataset and a set of rules, aggregations are performed
to produce for each country, possible date points (and
eventually other dimensions) a set of measure(s).
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
}
