% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_with_rules.R
\name{aggregate_with_rules}
\alias{aggregate_with_rules}
\title{Aggregate raw dataset from country rules}
\usage{
aggregate_with_rules(x)
}
\arguments{
\item{x}{raw dataset (an object of class \code{raw_level})}
}
\description{
From a raw dataset and a set of rules, aggregations are performed
to produce for each country and possible date points a set of three measures: CTY, CTA
and BZN.
}
\examples{
load_dir <- system.file(package = "antaDraft", "data_sample")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- aggregate_with_rules(load_data)
}
