% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_load_read.R
\name{anta_load_read}
\alias{anta_load_read}
\title{import load data from an entsoe repository}
\usage{
anta_load_read(data_dir = NULL)
}
\arguments{
\item{data_dir}{datasets directory}
}
\description{
import csv data representing load data
from an entsoe repository.
}
\examples{
if( dir.exists( Sys.getenv("LOAD_DIR") ) )
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
}
