% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anta_prod_filiere.R
\name{anta_prod_channel}
\alias{anta_prod_channel}
\title{import production data per channels from an entsoe repository}
\usage{
anta_prod_channel(data_dir = NULL, utf16 = FALSE)
}
\arguments{
\item{data_dir}{datasets directory}

\item{utf16}{whether text file are encoded in UTF16 (if not, UTF8)}
}
\description{
import csv data representing production per channels data
from an entsoe repository.
}
\examples{
if( dir.exists( Sys.getenv("CHANNEL_PROD_DIR") ) )
  load_data <- anta_prod_channel( data_dir = Sys.getenv("CHANNEL_PROD_DIR") )
}
