% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_learning_db.R
\name{as_learning_db}
\alias{as_learning_db}
\title{prepare an aggregated load dataset to be model}
\usage{
as_learning_db(dat, hour_decay = -1, keep = "summary")
}
\arguments{
\item{dat}{aggregated load dataset}

\item{hour_decay}{integer specifying to shift data Y from \code{hour_decay}
and add the resulting column as variable}
}
\description{
add columns to an aggregated load dataset
so that a model can be run.
}
