% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_learning_db.R
\name{as_learning_db}
\alias{as_learning_db}
\title{prepare an aggregated load dataset to be model}
\usage{
as_learning_db(data, target = "CTY")
}
\arguments{
\item{data}{aggregated load dataset}

\item{target}{column to manipulate when shifting or summarizing}
}
\description{
add columns to an aggregated load dataset
so that a model can be run.
}
\examples{
load_dir <- system.file(package = "antaDraft", "data_sample")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- aggregate_with_rules(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
as_learning_db(aggregated_db)
}
