% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_daily_mean.R
\name{augment_daily}
\alias{augment_daily}
\title{add daily average value}
\usage{
augment_daily(x, col, country_id = "country", decay = 1)
}
\arguments{
\item{x}{dataset}

\item{col}{column name of the measure to be aggregated and lagged}

\item{country_id}{column name of the country to be used in aggregations}

\item{decay}{lag as a number of days}
}
\description{
add daily average values to an existing dataset
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
aggregated_db <- augment_holiday(aggregated_db)
aggregated_db <- augment_seasons_id(aggregated_db)
aggregated_db <- augment_daylight(aggregated_db)
aggregated_db <- augment_daily(aggregated_db, col = "CTY", decay = 1)
aggregated_db <- augment_daily(aggregated_db, col = "CTY", decay = 2)
head(aggregated_db)
}
