% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_daylight.R
\name{augment_daylight}
\alias{augment_daylight}
\title{add day light columns}
\usage{
augment_daylight(x, loc_id = "country")
}
\arguments{
\item{x}{dataset}

\item{loc_id}{column name specifying the location column (country)}
}
\description{
add day light columns
}
\examples{
load_dir <- system.file(package = "antaDraft", "data_sample")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- aggregate_with_rules(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
aggregated_db <- augment_holiday(aggregated_db)
aggregated_db <- augment_seasons_id(aggregated_db)
aggregated_db <- augment_daylight(aggregated_db)
head(aggregated_db)
}
