% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_holiday.R
\name{augment_holiday}
\alias{augment_holiday}
\title{add holiday column}
\usage{
augment_holiday(x, country_id = "country")
}
\arguments{
\item{x}{dataset}

\item{country_id}{column name specifying the country column}
}
\description{
add holiday column
}
\examples{

if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
}

if( require(magrittr) )
  agg_db <- aggregate_with_rules(load_data) \%>\%
    augment_validation() \%>\%
    data_correct_with_rules() \%>\%
    augment_process_summary() \%>\%
    augment_holiday()
}
