% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_process_summary.R
\name{augment_process_summary}
\alias{augment_process_summary}
\title{add summary columns}
\usage{
augment_process_summary(data, colname = "summary")
}
\arguments{
\item{data}{dataset}

\item{colname}{name of the column to create that will contain
processed informations (i.e. original, invalid, corrected, ...)}
}
\description{
add validation column
}
\examples{
if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
  agg_db <- aggregate_with_rules(load_data)
  agg_db <- augment_validation(agg_db)
  agg_db <- data_correct_with_rules(agg_db)
  agg_db <- augment_process_summary(agg_db)
}
}
