% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_seasons_id.R
\name{augment_seasons_id}
\alias{augment_seasons_id}
\title{add weekly season id columns}
\usage{
augment_seasons_id(x, ts_key = "DateTime")
}
\arguments{
\item{x}{dataset}

\item{ts_key}{column name specifying the datetime column}
}
\description{
add weekly season id columns
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
aggregated_db <- augment_process_summary(aggregated_db)
aggregated_db <- augment_holiday(aggregated_db)
aggregated_db <- augment_seasons_id(aggregated_db)
head(aggregated_db)
}
