% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{augment_validation}
\alias{augment_validation}
\title{evaluate validation rules for each line of a dataset}
\usage{
augment_validation(data)
}
\arguments{
\item{data}{dataset}
}
\description{
add logical columns to a dataset. Each column is a test to
perform against dataset.
}
\examples{
load_dir <- system.file(package = "antaDraft", "data_sample")

load_data <- anta_load_read(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- aggregate_with_rules(load_data)
aggregated_db <- augment_validation(aggregated_db)
head(aggregated_db)
}
