% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{augment_validation}
\alias{augment_validation}
\title{evaluate validation rules for each line of a dataset}
\usage{
augment_validation(data)
}
\arguments{
\item{data}{dataset}
}
\description{
add logical columns to a dataset. Each column is a test to
perform against dataset.
}
\examples{
if( dir.exists( Sys.getenv("LOAD_DIR") ) ){
  load_data <- anta_load_read( data_dir = Sys.getenv("LOAD_DIR") )
  load_data <- augment_validation(data = load_data)
}
}
