% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{correct_db}
\alias{correct_db}
\title{correct a data frame}
\usage{
correct_db(db, rules)
}
\arguments{
\item{db}{data to be eventually modified}

\item{rules}{yaml file containing rules for corrections}
}
\description{
Correction of a data frame with a set of yaml rules
}
