% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_correct.R
\name{data_correct_with_rules}
\alias{data_correct_with_rules}
\title{correct datasets}
\usage{
data_correct_with_rules(data, refresh_validation = TRUE)
}
\arguments{
\item{data}{dataset}

\item{refresh_validation}{indicate to run \code{augment_validation}
after corrections.}
}
\description{
correct data based on condition expressed in a yaml file.
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")

load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(data = load_data)
head(load_data)

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)
aggregated_db <- data_correct_with_rules(aggregated_db)
head(aggregated_db)
}
