% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_raw_data.R
\name{fake_raw_data}
\alias{fake_raw_data}
\title{rebuild an expected fake dataset from raw data}
\usage{
fake_raw_data(raw_db, file_rules = NULL)
}
\arguments{
\item{raw_db}{the raw data returned by \code{read_load_files}}

\item{file_rules}{yaml file containing country rules}
}
\description{
rebuild an expected fake dataset from raw data
}
