% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_load_db.R
\name{fortify_load_db}
\alias{fortify_load_db}
\title{tidy load data}
\usage{
fortify_load_db(load_data)
}
\arguments{
\item{load_data}{data returned by \code{\link{import_load_db}}}
}
\description{
tidy and correct load data
}
