% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{fortify_qualcon}
\alias{fortify_qualcon}
\title{eval validation rules against a dataset}
\usage{
fortify_qualcon(dat)
}
\arguments{
\item{db}{data to be confronted with rules}
}
\description{
Confront data with a set of validation rules
}
