% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{fortify_qualcon}
\alias{fortify_qualcon}
\title{eval validation rules against a dataset}
\usage{
fortify_qualcon(dat)
}
\arguments{
\item{dat}{errors data.frame to be summarised. It should be a
data.frame returned by the call to \code{\link{qualcon}}}
}
\description{
Confront data with a set of validation rules
}
\examples{
library(antadraft)
rep_path <- system.file(package = "antadraft", "files/load")
load_db <- read_load_files(rep_path)
db <- fortify_from_rules(raw_db = load_db)
db_errors <- qualcon(db)
errors_summary <- fortify_qualcon(db_errors)
}
