% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_antares_project.R
\name{init_antares_project}
\alias{init_antares_project}
\title{create an empty antares project}
\usage{
init_antares_project(path)
}
\arguments{
\item{path}{folder to be created}
}
\description{
create an empty antares project to be then
enriched with data.
}
\examples{
init_antares_project("demo_proj")
}
