% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_compare.R
\name{prod_compare}
\alias{prod_compare}
\title{compare production by types and production by groups}
\usage{
prod_compare(prod_type, group_prod)
}
\arguments{
\item{prod_type}{data produced by \code{anta_prod_type}}

\item{group_prod}{data produced by \code{anta_prod_group}}
}
\description{
compare by types and by groups production data.
}
\examples{
production_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B01")
capacity_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B06")
prod_by_types <- anta_prod_type(production_dir, capacity_dir)
production_dir <- system.file(package = "antaDraft", "data_sample",
  "prod_sample_20160129/B02")
prod_by_groups <- anta_prod_group(production_dir)
comparison <- prod_compare( prod_by_types, prod_by_groups )
comparison <- augment_validation(comparison)
qc_comp <- qualcon(comparison)
render_quality(qc_comp, "qc_cmp_dir")
}
