% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_compare.R
\name{prod_compare}
\alias{prod_compare}
\title{compare production by types and production by groups}
\usage{
prod_compare(prod_type, group_prod)
}
\arguments{
\item{prod_type}{data produced by \code{anta_prod_type}}

\item{group_prod}{data produced by \code{anta_prod_group}}
}
\description{
compare by types and by groups production data.
}
