% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{qualcon}
\alias{qualcon}
\title{eval validation rules against a dataset}
\usage{
qualcon(db, rules = system.file(package = "antadraft",
  "yaml_data/validate/validation_rules.yml"), fp_rules = system.file(package =
  "antadraft", "yaml_data/validate/false_positive_rules.yml"))
}
\arguments{
\item{db}{data to be confronted with rules}

\item{rules}{yaml file containing rules to be used to validate data}

\item{fp_rules}{yaml file containing false positive rules to exclude
a posteriori.}
}
\description{
Confront data with a set of validation rules
}
\examples{
library(antadraft)
rep_path <- system.file(package = "antadraft", "files/load")
load_db <- read_load_files(rep_path)
db <- fortify_from_rules(raw_db = load_db)
db_errors <- qualcon(db)
}
