% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{qualcon}
\alias{qualcon}
\title{eval validation rules against a dataset}
\usage{
qualcon(db, rules = system.file(package = "antadraft",
  "validation_rules.yml"), fp_rules = system.file(package = "antadraft",
  "false_positive_rules.yml"))
}
\arguments{
\item{db}{data to be confronted with rules}

\item{rules}{yaml file containing rules to evaluate}

\item{fp_rules}{yaml file containing false positive rules to exclude
a posteriori.}
}
\description{
Confront data with a set of validation rules
}
