% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{qualcon}
\alias{qualcon}
\alias{qualcon.controled}
\title{perform quality control}
\usage{
qualcon(x)

\method{qualcon}{controled}(x)
}
\arguments{
\item{x}{data where controls have been evaluated}
}
\description{
perform quality control
}
