% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{qualcon}
\alias{qualcon}
\title{eval validation rules against a dataset}
\usage{
qualcon(db, yaml_rules = system.file(package = "antadraft",
  "validation_rules.yml"))
}
\arguments{
\item{db}{data to be confronted with rules}
}
\description{
Confront data with a set of validation rules
}
