% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{render_quality}
\alias{render_quality}
\title{report QC results}
\usage{
render_quality(x, dir)
}
\arguments{
\item{x}{an object of class \code{qualcon}}

\item{dir}{directory where reports should be written}
}
\description{
create reports from results of \code{qualcon} call.
}
