% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{render_quality}
\alias{render_quality}
\alias{render_quality.raw_level}
\alias{render_quality.aggregated}
\alias{render_quality.raw_prod_type}
\alias{render_quality.aggregated_prod}
\alias{render_quality.raw_prod_group}
\alias{render_quality.qualcon_comp_prod}
\title{report QC results}
\usage{
render_quality(x, dir)

\method{render_quality}{raw_level}(x, dir)

\method{render_quality}{aggregated}(x, dir)

\method{render_quality}{raw_prod_type}(x, dir)

\method{render_quality}{aggregated_prod}(x, dir)

\method{render_quality}{raw_prod_group}(x, dir)

\method{render_quality}{qualcon_comp_prod}(x, dir)
}
\arguments{
\item{x}{an object of class \code{qualcon}}

\item{dir}{directory where reports should be written}
}
\description{
create reports from results of \code{qualcon} call.
}
\examples{
load_dir <- system.file(package = "antaDraft",
  "data_sample/load_sample_2017")
load_data <- anta_load(data_dir = load_dir )
load_data <- augment_validation(load_data)

qc <- qualcon(load_data)
render_quality(qc, "raw_qc")

aggregated_db <- agg_data(load_data)
aggregated_db <- augment_validation(aggregated_db)

qc <- qualcon(aggregated_db)
render_quality(qc, "agg_qc")
}
