% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcon.R
\name{report_errors_summary}
\alias{report_errors_summary}
\title{produce reports for each country and error type}
\usage{
report_errors_summary(error_sum, yaml_v_rules = NULL, dir = getwd())
}
\arguments{
\item{error_sum}{result returned by function \code{fortify_qualcon}}

\item{yaml_v_rules}{yaml file containing rules to be used to validate data}

\item{dir}{directory where to write reports}
}
\description{
produce reports for each country and error type
}
